+++ ext/types/types.go
@@ -1,0 +1,2 @@
+// Package types provides a list of scalar and structure types.
+package types